part-3/backend/requirements.txt
# Flask - Web framework
Flask==3.0.0

# Flask-CORS - Allow frontend to call backend API
flask-cors==4.0.0

# Flask-SQLAlchemy - Database ORM
Flask-SQLAlchemy==3.1.1

# bcrypt - Password hashing
bcrypt==4.1.2

# PyJWT - Create and verify JWT tokens
PyJWT==2.8.0
